+++ src/lib/components/prompt/PromptBar.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
+	import { getPrompt } from '$lib/api';
 	import type { Tables } from '$lib/supabase';
-	import { ChevronDown, ChevronUp } from 'lucide-svelte';
+	import { tooltip } from '$lib/tooltip.svelte';
+	import { ChevronDown, ChevronUp, MoveLeft, MoveRight } from 'lucide-svelte';
 	import PromptOptions from '../options/PromptOptions.svelte';
 	import PromptEditor from './PromptEditor.svelte';
 	import PromptSuggestions from './PromptSuggestions.svelte';
@@ -36,13 +38,42 @@
 		setPrompt: () => void;
 		loadPrompt: (id: number | null) => void;
 	}>();
+
+	async function forward() {
+		let target = forwardTarget;
+		while (target.parent_prompt_id !== null) {
+			if (target.parent_prompt_id === prompt.id) {
+				loadPrompt(target.id);
+				return;
+			}
+			target = await getPrompt(target.parent_prompt_id);
+		}
+	}
 </script>
 
 <div class="flex h-full w-[450px] shrink-0 flex-col border-r px-6 py-4">
 	<div class="mb-2 flex items-end justify-between">
-		<h1>Prompt</h1>
+		<div class="flex items-center gap-4">
+			<h1 class="mr-2">Prompt</h1>
+			<button
+				class="group h-5 w-5 cursor-pointer text-gray-500 transition-colors hover:text-gray-900 disabled:cursor-not-allowed disabled:text-gray-300"
+				disabled={prompt.parent_prompt_id === null}
+				onclick={() => loadPrompt(prompt.parent_prompt_id)}
+				use:tooltip={{ text: 'Go to previous prompt' }}
+			>
+				<MoveLeft />
+			</button>
+			<button
+				class="group h-5 w-5 cursor-pointer text-gray-500 transition-colors hover:text-gray-900 disabled:cursor-not-allowed disabled:text-gray-300"
+				disabled={prompt.id === forwardTarget.id}
+				onclick={forward}
+				use:tooltip={{ text: 'Go to next prompt' }}
+			>
+				<MoveRight />
+			</button>
+		</div>
 		<button
-			class="flex items-center gap-1 opacity-40 transition-all hover:opacity-100"
+			class="flex h-full items-center gap-1 opacity-40 transition-all hover:opacity-100"
 			onclick={() => (showOptions = !showOptions)}
 		>
 			<span class="text-black">Model Options</span>
@@ -60,12 +91,10 @@
 		<PromptEditor
 			{prompt}
 			{hoveredSuggestion}
-			{forwardTarget}
 			{setPrompt}
 			bind:suggestionApplied
 			bind:editedPrompt
 			{setPromptMaximized}
-			{loadPrompt}
 		/>
 	</div>
 	{#if projectId}
